# optimize extract performance