--- a/src/extract.rs
+++ b/src/extract.rs
@@ -1,4 +1,3 @@
-use std::error::Error;
 use std::fs::File;
 use std::fs::*;
 use std::io::BufReader;
@@ -6,39 +5,32 @@ use std::io::{Cursor, Read, Seek, SeekFrom, Write};
 use std::path::PathBuf;
 use std::*;
 
-use flate2::read::ZlibDecoder;
-
 use rayon::prelude::*;
 
 use crate::ls;
 use crate::rf;
 use ls::LSEntry;
 
-pub use ls::ls_str::crc32;
-
-pub struct DecompressLater {
+pub struct FileData {
     pub path: PathBuf,
     pub data: Vec<u8>,
     pub cmp: bool,
 }
 
-fn decode_reader(bytes: &[u8]) -> Result<Vec<u8>, Box<dyn Error>> {
-    let mut z = ZlibDecoder::new(bytes);
-    let mut s: Vec<u8> = Vec::new();
-    z.read_to_end(&mut s)?;
-    Ok(s)
+fn decode_zlib(bytes: &[u8]) -> Result<Vec<u8>, zune_inflate::errors::InflateDecodeErrors> {
+    zune_inflate::DeflateDecoder::new(bytes).decode_zlib()
 }
 
-pub fn extract(_ls_file: PathBuf, _dt_file: PathBuf, _dt1_file: PathBuf, _out_folder: PathBuf) {
-    let ls = ls::LSFile::open(_ls_file).unwrap();
+pub fn extract(ls_file: PathBuf, dt_file: PathBuf, dt1_file: PathBuf, out_folder: PathBuf) {
+    let ls = ls::LSFile::open(ls_file).unwrap();
 
     let rf_file_info = ls.find("resource");
 
     let mut dts = Vec::new();
-    dts.push(BufReader::new(File::open(_dt_file).unwrap()));
-    if !_dt1_file.as_os_str().is_empty() {
-        println!("{0}", _dt1_file.as_os_str().to_str().unwrap());
-        dts.push(BufReader::new(File::open(_dt1_file).unwrap()));
+    dts.push(BufReader::new(File::open(dt_file).unwrap()));
+    if !dt1_file.as_os_str().is_empty() {
+        println!("{0}", dt1_file.as_os_str().to_str().unwrap());
+        dts.push(BufReader::new(File::open(dt1_file).unwrap()));
     }
     let mut dt_idx;
     match rf_file_info.dt_index {
@@ -49,18 +41,20 @@ pub fn extract(_ls_file: PathBuf, _dt_file: PathBuf, _dt1_file: PathBuf, _out_fo
         .seek(SeekFrom::Start(rf_file_info.offset as u64))
         .unwrap();
 
+    // Alocate memory for rf
     let mut rf_data = vec![0u8; rf_file_info.size as usize];
-    //Alocate memory for rf
     dts[dt_idx as usize].read_exact(&mut rf_data).unwrap();
     let mut filetest = File::create("testOut.rf").unwrap();
     filetest.write_all(&rf_data).unwrap();
-    //Read from buffer into that alocated memory
+
+    // Read from buffer into that allocated memory
     let mut rf_cursor = Cursor::new(&rf_data);
     let rf = rf::RFFile::read(&mut rf_cursor);
     filetest.seek(SeekFrom::Start(0x80)).unwrap();
     filetest.write_all(&rf.debug_extract).unwrap();
+
     let mut stringbuild: PathBuf = PathBuf::new();
-    let mut path_out: Vec<PathBuf> = Vec::new();
+    let mut out_paths: Vec<PathBuf> = Vec::new();
     let mut oneshot = false;
     for n in &rf.entrys {
         while n.folder_depth < stringbuild.components().count() as u32 + 1 && oneshot {
@@ -68,19 +62,25 @@ pub fn extract(_ls_file: PathBuf, _dt_file: PathBuf, _dt1_file: PathBuf, _out_fo
         }
         oneshot = true;
         stringbuild.push(PathBuf::from(&n.file_name));
-        path_out.push(stringbuild.iter().collect());
+        out_paths.push(stringbuild.iter().collect());
     }
+
     let mut lsoffset: LSEntry = LSEntry::default();
-    let mut dolater: Vec<DecompressLater> = Vec::new();
-    for n in 0..rf.entrys.len() {
-        let mut cur_data = path_out[n].clone();
-        let mut folder_path = _out_folder.join(&path_out[n]);
-        if rf.entrys[n].is_folder && !folder_path.exists() {
-            create_dir_all(&folder_path).unwrap();
+    let mut files_to_write: Vec<FileData> = Vec::new();
+
+    // Create all necessary folders before extracting files.
+    for (rf_entry, out_path) in rf.entrys.iter().zip(&out_paths) {
+        let path = out_folder.join(out_path);
+        if rf_entry.is_folder && !path.exists() {
+            create_dir_all(&path).unwrap();
         }
-        if rf.entrys[n].is_compressed {
-            if rf.entrys[n].is_folder {
-                folder_path.push(PathBuf::from("packed"));
+    }
+
+    for (rf_entry, out_path) in rf.entrys.iter().zip(out_paths) {
+        let path = out_folder.join(&out_path);
+        if rf_entry.is_compressed {
+            let mut cur_data = out_path.clone();
+            if rf_entry.is_folder {
                 cur_data.push(PathBuf::from("packed"));
             };
             lsoffset = ls.find(&format!(
@@ -91,41 +91,33 @@ pub fn extract(_ls_file: PathBuf, _dt_file: PathBuf, _dt1_file: PathBuf, _out_fo
                 Some(x) => dt_idx = x,
                 None => dt_idx = 0,
             }
-            let mut cur_data = vec![0u8; lsoffset.size as usize];
-            dts[dt_idx as usize]
-                .seek(SeekFrom::Start(lsoffset.offset as u64))
-                .unwrap();
-            dts[dt_idx as usize].read_exact(&mut cur_data).unwrap();
-            let _fs_cursor = Cursor::new(&cur_data);
-            std::fs::write(&folder_path, cur_data).unwrap();
-        } else if !rf.entrys[n].is_folder {
-            let mut cur_cmp_data = vec![0u8; rf.entrys[n].file_size_cmp as usize];
+        } else if !rf_entry.is_folder {
+            let mut cur_cmp_data = vec![0u8; rf_entry.file_size_cmp as usize];
             dts[dt_idx as usize]
                 .seek(SeekFrom::Start(
-                    (lsoffset.offset + rf.entrys[n].file_offset) as u64,
+                    (lsoffset.offset + rf_entry.file_offset) as u64,
                 ))
                 .unwrap();
             dts[dt_idx as usize].read_exact(&mut cur_cmp_data).unwrap();
-            std::fs::write(
-                &folder_path,
-                if rf.entrys[n].file_size_cmp != rf.entrys[n].file_size {
-                    decode_reader(&cur_cmp_data[..]).unwrap()
-                } else {
-                    cur_cmp_data
-                },
-            )
-            .unwrap();
+
+            files_to_write.push(FileData {
+                path,
+                data: cur_cmp_data,
+                cmp: rf_entry.file_size_cmp != rf_entry.file_size,
+            });
         }
     }
-    dolater.par_iter().for_each(|c| {
-        std::fs::write(
+
+    files_to_write.par_iter().for_each(|c| {
+        if let Err(e) = std::fs::write(
             &c.path,
             if c.cmp {
-                decode_reader(&c.data[..]).unwrap()
+                decode_zlib(&c.data[..]).unwrap()
             } else {
                 c.data.clone()
             },
-        )
-        .unwrap()
+        ) {
+            println!("Failed to extract {:?}: {e}", &c.path)
+        }
     });
 }