… Cargo.toml
… src/lib.rs
… src/ls/ls_str.rs