This PR uses a faster zlib decoder, multithreading, and better release options to improve the extraction performance. All testing was done on Windows 11 with a Ryzen 3700x and Samsung 980 Pro with Windows defender temporarily disabled. 

zune inflate 59s -> 39s 
relase profile 39s -> 34s
rayon 34s -> 10s

This also skips generating the "packed" file in each folder and continues extracting even if a file fails. It looks like `fighter/rockman/model/plantbarrier/c00/material_anime/eat/tFlip.mta` is the only error on Smash 4 Wii U.