@smb123w64gb 2025-02-04
Looks good. Nice to see someone clean up some of my early Rust code! 
Thank you